requirements.txt
# requirements.txt

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
opencv-python>=4.8.0
Pillow>=9.5.0

# Optional (EfficientNet)
efficientnet-pytorch>=0.7.1

# Testing
pytest>=7.4.0

# Config
PyYAML>=6.0

# Progress bar
tqdm>=4.65.0
